# docs/swagger.yaml
definitions:
...
  dto.ResponseJson:
    properties:
      data: {}
      error:
        type: string
      limit:
        type: integer
      message:
        type: string
      offset:
        type: integer
      total_records:
        type: integer
    type: object
  models.Category:
    properties:
      category_id:
        type: string
      category_name:
        type: string
      created_at:
        type: string
      description:
        type: string
      posts:
        items:
          $ref: '#/definitions/models.Post'
        type: array
      updated_at:
        type: string
    type: object
  models.Comment:
    properties:
      comment_id:
        type: string
      content:
        type: string
      created_at:
        type: string
      post_id:
        type: string
      replies:
        items:
          $ref: '#/definitions/models.Reply'
        type: array
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  models.Post:
    properties:
      category_id:
        type: string
      comments:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      content:
        type: string
      created_at:
        type: string
      description:
        type: string
      post_id:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  models.Reply:
    properties:
      comment_id:
        type: string
      content:
        type: string
      created_at:
        type: string
      reply_id:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
host: localhost:5030
info:
  contact:
    email: rsi28c@gmail.com
    name: admin
    url: https://github.com/marees7/rishi-aug-2024.git
  description: This is a blog post api where users can create posts and add comments
    for it
  title: Blog posts API
  version: 0.0.1
paths:
...
  /v1/admin/categories:
    post:
      consumes:
      - application/json
      description: Create a new category
      operationId: create-category
      parameters:
      - description: Enter the category details that need to be created
        in: body
        name: CreateDetails
        required: true
        schema:
          $ref: '#/definitions/models.Category'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ResponseJson'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ResponseJson'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ResponseJson'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ResponseJson'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ResponseJson'
      security:
      - JWT: []
      summary: Create categories
      tags:
      - Category
  /v1/admin/categories/{categoryID}:
    delete:
      consumes:
      - application/json
      description: Delete a existing category
      operationId: Delete-category
      parameters:
      - description: Enter the category id
        in: path
        name: categoryID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponseJson'
        "304":
          description: Not Modified
          schema:
            $ref: '#/definitions/dto.ResponseJson'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ResponseJson'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ResponseJson'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ResponseJson'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ResponseJson'
      security:
      - JWT: []
      summary: Delete categories
      tags:
      - Category
    put:
      consumes:
      - application/json
      description: Update a existing category
      operationId: update-category
      parameters:
      - description: Enter the category id
        in: path
        name: categoryID
        required: true
        type: string
      - description: Enter the changes need to be done
        in: body
        name: UpdateDetails
        required: true
        schema:
          $ref: '#/definitions/models.Category'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponseJson'
        "304":
          description: Not Modified
          schema:
            $ref: '#/definitions/dto.ResponseJson'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ResponseJson'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ResponseJson'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ResponseJson'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ResponseJson'
      security:
      - JWT: []
      summary: Update categories
      tags:
      - Category
...
  /v1/users/categories:
    get:
      description: Get all the available categories
      operationId: get-category
      parameters:
      - description: Enter the limit
        in: query
        name: limit
        type: string
      - description: Enter the offset
        in: query
        name: offset
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponseJson'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ResponseJson'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ResponseJson'
      security:
      - JWT: []
      summary: Get categories
      tags:
      - Category
...
swagger: "2.0"